// src/Interfaces/HistorialDetalle.tsx
export interface HistorialDetalleInterface {
    $id: string
    IdGastoViajeDetalle: number
    Tipo: string
    categoria: string
    UsuarioAsesor: string
    Proveedor: string
    NoFactura: string
    Descripcion: string
    DescripcionAdmin: string
    importeExento: number
    importeGravado: number
    ValorFactura: number
    FechaFactura: string
    FechaCreacion: string
    imagen: string
    Admin: string
    serie: string
  }